+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   support for BigQuery `select as struct '1' as bb, 2 as aa` syntax
 - Fix bug [#1079](https://github.com/sqlfluff/sqlfluff/issues/1079), addressing
   issues with L025 and L026 with BigQuery column references involving `STRUCT`
+- Fix bug [#1080](https://github.com/sqlfluff/sqlfluff/issues/1080), add
+  SET SCHEMA and DROP SCHEMA support to ANSI dialect.
 
 Contributors:
 - [@GitHub-Username](Link to GitHub profile) ([#PR-Number](Link to PR))